minimize logging for refactor
