+++ packages/sdk/src/realtime/livekit-connection.ts
@@ -11,7 +11,6 @@
 
 import {
   ConnectionState as LKConnectionState,
-  type LocalTrack,
   type RemoteParticipant,
   type RemoteTrack,
   Room,
@@ -20,14 +19,10 @@ import {
   TrackEvent,
   type TrackPublishOptions,
 } from "livekit-client";
-
-const INFERENCE_SERVER_IDENTITY_PREFIX = "inference-server-";
 import mitt from "mitt";
 
-import type { Logger } from "../utils/logger";
 import { buildUserAgent } from "../utils/user-agent";
 import type { RealtimeObservability } from "./observability/realtime-observability";
-import type { StatsProvider } from "./observability/webrtc-stats";
 import type {
   ConnectionChangeDetails,
   ConnectionState,
@@ -41,6 +36,8 @@ import type {
   SessionIdMessage,
   SetImageAckMessage,
 } from "./types";
+
+const INFERENCE_SERVER_IDENTITY_PREFIX = "inference-server-";
 
 const SETUP_TIMEOUT_MS = 30_000;
 const ROOM_INFO_TIMEOUT_MS = 15_000;
@@ -94,7 +91,6 @@ interface LiveKitCallbacks {
   onError?: (error: Error) => void;
   initialImage?: string;
   initialPrompt?: { text: string; enhance?: boolean };
-  logger?: Logger;
   observability?: RealtimeObservability;
 }
 
@@ -112,28 +108,11 @@ export class LiveKitConnection {
   private room: Room | null = null;
   private localStream: MediaStream | null = null;
   private connectionReject: ((error: Error) => void) | null = null;
-  private logger: Logger;
-  private statsProvider: StatsProvider | null = null;
   private remoteStream: MediaStream | null = null;
-  private wsOpenedAt: number | null = null;
-  private lastServerError: string | null = null;
   state: ConnectionState = "disconnected";
   websocketMessagesEmitter = mitt<WsMessageEvents>();
 
-  constructor(private callbacks: LiveKitCallbacks = {}) {
-    this.logger = callbacks.logger ?? { debug() {}, info() {}, warn() {}, error() {} };
-  }
-
-  /**
-   * Stats provider for the LiveKit connection. Aggregates
-   * `track.getRTCStatsReport()` from every local (outbound) and remote
-   * (inbound) track in the room into a single RTCStatsReport-compatible
-   * Map. That's the minimum surface WebRTCStatsCollector.parse() needs —
-   * it calls `.forEach` and keys off `report.type`.
-   */
-  getStatsProvider(): StatsProvider | null {
-    return this.statsProvider;
-  }
+  constructor(private callbacks: LiveKitCallbacks = {}) {}
 
   async connect(url: string, localStream: MediaStream | null, timeout: number, integration?: string): Promise<void> {
     this.localStream = localStream;
@@ -152,30 +131,17 @@ export class LiveKitConnection {
 
     try {
       // Phase 1 — control WS + livekit_join/livekit_room_info handshake.
-      const roomInfoStart = performance.now();
       await Promise.race([this.openControlWs(wsUrl, timeout), connectAbort]);
       const roomInfo = await Promise.race([this.requestRoomInfo(), connectAbort]);
       this.setState("connecting");
-      this.callbacks.observability?.diagnostic("phaseTiming", {
-        phase: "websocket",
-        durationMs: performance.now() - roomInfoStart,
-        success: true,
-      });
 
       // Phase 2 — join the SFU room and publish local tracks.
-      const roomStart = performance.now();
       this.room = new Room(LIVEKIT_ROOM_OPTIONS);
       this.room.prepareConnection(roomInfo.livekit_url, roomInfo.token).catch(() => {});
       await Promise.race([this.joinRoom(roomInfo), connectAbort]);
-      this.callbacks.observability?.diagnostic("phaseTiming", {
-        phase: "webrtc-handshake",
-        durationMs: performance.now() - roomStart,
-        success: true,
-      });
 
       // Phase 3 — optional initial conditioning over the control WS.
       if (this.callbacks.initialImage) {
-        const imageStart = performance.now();
         await Promise.race([
           this.setImageBase64(this.callbacks.initialImage, {
             prompt: this.callbacks.initialPrompt?.text,
@@ -183,27 +149,10 @@ export class LiveKitConnection {
           }),
           connectAbort,
         ]);
-        this.callbacks.observability?.diagnostic("phaseTiming", {
-          phase: "initial-image",
-          durationMs: performance.now() - imageStart,
-          success: true,
-        });
       } else if (this.callbacks.initialPrompt) {
-        const promptStart = performance.now();
         await Promise.race([this.sendInitialPrompt(this.callbacks.initialPrompt), connectAbort]);
-        this.callbacks.observability?.diagnostic("phaseTiming", {
-          phase: "initial-prompt",
-          durationMs: performance.now() - promptStart,
-          success: true,
-        });
       } else if (localStream) {
-        const passthroughStart = performance.now();
         await Promise.race([this.setImageBase64(null, { prompt: null }), connectAbort]);
-        this.callbacks.observability?.diagnostic("phaseTiming", {
-          phase: "initial-prompt",
-          durationMs: performance.now() - passthroughStart,
-          success: true,
-        });
       }
 
       this.setState("connected");
@@ -269,8 +218,7 @@ export class LiveKitConnection {
       this.room.disconnect().catch(() => {});
       this.room = null;
     }
-    this.statsProvider = null;
-    this.callbacks.observability?.setStatsProvider(null);
+    this.callbacks.observability?.setLiveKitRoom(null);
     this.remoteStream = null;
     if (this.ws) {
       try {
@@ -289,26 +237,13 @@ export class LiveKitConnection {
 
   private async openControlWs(wsUrl: string, timeout: number): Promise<void> {
     await new Promise<void>((resolve, reject) => {
-      const openStart = performance.now();
       const timer = setTimeout(() => reject(new Error("WebSocket timeout")), timeout);
-      this.lastServerError = null;
       this.ws = new WebSocket(wsUrl);
-      this.wsOpenedAt = openStart;
       this.ws.onopen = () => {
         clearTimeout(timer);
         resolve();
       };
       this.ws.onclose = (e) => {
-        const details = {
-          code: e.code,
-          reason: e.reason || this.lastServerError || "(none)",
-          wasClean: e.wasClean,
-          serverError: this.lastServerError,
-          uptimeMs: this.wsOpenedAt ? Math.round(performance.now() - this.wsOpenedAt) : null,
-        };
-        if (this.state !== "disconnected" || this.lastServerError) {
-          this.logger.warn("LiveKit control WS closed unexpectedly", details);
-        }
         // If the room is still connecting this also aborts the connect flow.
         this.connectionReject?.(new Error(`WebSocket closed: ${e.code} ${e.reason}`));
         if (!this.room || this.room.state !== LKConnectionState.Connected) {
@@ -321,11 +256,8 @@ export class LiveKitConnection {
       this.ws.onmessage = (e) => {
         try {
           this.handleControlMessage(JSON.parse(e.data));
-        } catch (error) {
-          this.logger.error("LiveKit control WS message parse error", {
-            error: String(error),
-            preview: typeof e.data === "string" ? e.data.slice(0, 200) : "(non-string)",
-          });
+        } catch {
+          // Ignore malformed control messages; valid server errors use the `error` message type.
         }
       };
     });
@@ -390,13 +322,9 @@ export class LiveKitConnection {
         this.websocketMessagesEmitter.emit("generationTick", msg);
         break;
       case "generation_ended":
-        if (msg.reason && msg.reason !== "disconnect") {
-          this.lastServerError = msg.reason;
-        }
         this.websocketMessagesEmitter.emit("generationEnded", msg);
         break;
       case "error": {
-        this.lastServerError = msg.error ?? null;
         const error = new Error(msg.error) as Error & { source?: string };
         error.source = "server";
         this.callbacks.onError?.(error);
@@ -437,11 +365,7 @@ export class LiveKitConnection {
 
     await this.room.connect(info.livekit_url, info.token);
 
-    // Wire up the stats provider now that the room has local+remote
-    // participant objects available. Observability owns polling and
-    // replaces the collector when reconnect creates a new room/provider.
-    this.statsProvider = createLiveKitStatsProvider(this.room);
-    this.callbacks.observability?.setStatsProvider(this.statsProvider);
+    this.callbacks.observability?.setLiveKitRoom(this.room);
 
     if (this.localStream) {
       await this.publishLocalTracks(this.localStream);
@@ -501,48 +425,4 @@ export class LiveKitConnection {
       this.callbacks.onStateChange?.(state, details);
     }
   }
-}
-
-function createLiveKitStatsProvider(room: Room): StatsProvider {
-  let uid = 0;
-
-  const collectFromTrack = async (
-    track: LocalTrack | RemoteTrack | undefined,
-    entries: Array<[string, unknown]>,
-  ): Promise<void> => {
-    if (!track) return;
-    let report: RTCStatsReport | undefined;
-    try {
-      report = await track.getRTCStatsReport();
-    } catch {
-      // Track is likely unmuted/unattached or the PC is mid-teardown — skip it.
-      return;
-    }
-    if (!report) return;
-    report.forEach((stat, id) => {
-      entries.push([`${id}#${uid++}`, stat]);
-    });
-  };
-
-  return {
-    async getStats(): Promise<RTCStatsReport> {
-      const entries: Array<[string, unknown]> = [];
-
-      // Outbound: the local participant's published tracks (what we send to
-      // the SFU — the server reads these as its inbound video/audio).
-      for (const pub of room.localParticipant.trackPublications.values()) {
-        await collectFromTrack(pub.track, entries);
-      }
-
-      for (const participant of room.remoteParticipants.values()) {
-        for (const pub of participant.trackPublications.values()) {
-          await collectFromTrack(pub.track as RemoteTrack | undefined, entries);
-        }
-      }
-
-      // `Map` is structurally compatible with `RTCStatsReport` for the
-      // callers we care about (WebRTCStatsCollector.parse uses forEach).
-      return new Map(entries) as unknown as RTCStatsReport;
-    },
-  };
 }
